requirements.txt
arrow==1.3.0
asgiref==3.7.2
binaryornot==0.4.4
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.5.0
Django==4.2.7
django-browser-reload==1.12.1
django-cors-headers==4.3.1
django-tailwind @ git+https://github.com/timonweb/django-tailwind.git@4ba910732cd3782b314ef3a13ce7bdc47c764ffd
djangorestframework==3.14.0
idna==3.6
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mysqlclient==2.2.1
Pillow==10.1.0
Pygments==2.17.2
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.4
PyYAML==6.0.1
requests==2.31.0
require==0.1.1
rich==13.7.0
six==1.16.0
sqlparse==0.4.4
tailwind==3.1.5b0
text-unidecode==1.3
types-python-dateutil==2.8.19.14
tzdata==2023.3
urllib3==2.1.0
whitenoise==6.6.0
